test(skills): split test_skills.py into topic modules (closes #331)
The file was 670 LOC with 17 test methods spanning two distinct
concerns. Split for readability and to bring it under the 500 LOC
oversized threshold flagged by tech-lead's audit.

New layout (each file < 500 LOC):

- _skill_helpers.py             — shared discovery + parsing helpers
- test_skill_invariants.py      — file-level rules (footer, catalog sync)
- test_agent_frontmatter.py     — per-agent frontmatter rules
                                  (tick, output, scope contract,
                                  custom-doc, init, pilot receipts,
                                  registry alignment, …)

test_skills.py becomes a thin aggregator that imports both classes so
unittest.main() discovers them. CI's
`python3 claude-skills/tests/test_skills.py` invocation keeps working
unchanged. Each topic module is also independently runnable.

Test count is unchanged (17 → 17 across the three modules).
